add validation of user input

+++ SpriteSortingPlugin/Assets/SpriteSortingPlugin/Editor/Survey/Data/UsabilityData.cs
@@ -10,7 +10,8 @@ namespace SpriteSortingPlugin.Survey.Data
         public string highlights = "";
         public string lowlights = "";
 
-        public int[] ratingAnswers = Array.ConvertAll(new int[3], i => 50);
+        //TODO float or int?
+        public float[] ratingAnswers = Array.ConvertAll(new float[3], i => 50f);
 
         public string missingCriteriaText = "";
